Made BottomSheetNav a custom component
This reduces code complexity within Staff & Patient Profile and allows for easy debugging / changing later on.
